requirements.txt
# Core Aptos dependencies
aptos-sdk>=0.11.0
python-telegram-bot>=20.0
aiosqlite>=0.19.0

# Data analysis and ML for trading strategies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
ta>=0.10.0

# Async and networking
aiohttp>=3.8.0
websockets>=10.0
requests>=2.25.0

# Cryptography and security
cryptography>=3.4.0

# Additional dependencies for Aptos integration
pynacl>=1.5.0
ed25519>=1.5
bcs>=0.1.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=22.0.0
flake8>=4.0.0

# Optional: For enhanced functionality
python-dotenv>=0.19.0
schedule>=1.2.0
colorlog>=6.0.0
